pyproject.toml
[build-system]
requires = ["hatchling >= 1.27.0"]
build-backend = "hatchling.build"

[project]
version = "0.0.1"
name = "pip-security-worker"
description = "The worker for the pip-security eco-system."
requires-python = ">=3.13"
dependencies = [
    "defusedxml>=0.7.1",
    "kafka-python>=2.2.7",
    "neo4j>=5.28.1",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "sentry-sdk>=2.29.0",
]
authors = [
    { name = "Peter McDonald", email = "git@petermcdonald.co.uk" },
]
maintainers = [
    { name = "Peter McDonald", email = "git@petermcdonald.co.uk" },
]
readme = {file = "README.md", content-type = "text/markdown"}
license = "MIT"
license-files = ["LICENSE"]
keywords = ["pip", "security"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.13",
]

[project.urls]
Homepage = "https://github.com/petermcd/PIP-Security-Worker"
Documentation = "https://github.com/petermcd/PIP-Security-Worker"
Repository = "https://github.com/petermcd/PIP-Security-Worker.git"
Issues = "https://github.com/petermcd/PIP-Security-Worker/issues"

[tool.hatch.build.targets.wheel]
packages = ["pip_security_worker"]

[project.scripts]
analyze = "pip_security_worker.cli:run_analysis"
update_advisory_db = "pip_security_worker.cli:run_update_advisory_db"
updated_packages = "pip_security_worker.cli:run_recent_updated_packages"

[tool.ruff]

line-length = 120
target-version = "py313"
indent-width = 4

[tool.ruff.lint.isort]
combine-as-imports = true

[tool.ruff.lint]
fixable = ["ALL"]

[tool.ruff.format]
quote-style = "single"

[tool.mypt.overrides]
[[tool.mypy.overrides]]
module = ["dotenv.*", "kafka.*", "neo4j.*", "pytest.*", "sentry_sdk.*"]
ignore_missing_imports = true

[dependency-groups]
dev = [
    "pytest-cov>=6.1.1",
    "pytest-mock>=3.14.0",
]
